+++ src/main/java/com/julvez/pfc/teachonsnap/controller/common/FollowAuthorController.java
@@ -32,7 +32,7 @@ public class FollowAuthorController extends CommonController {
 
 	private StringManager stringManager = StringManagerFactory.getManager();
 	
-	private final int MAX_RESULTS_PAGE = properties.getNumericProperty(UserPropertyName.MAX_PAGE_RESULTS);
+	private final int MAX_RESULTS_PAGE = (int)properties.getNumericProperty(UserPropertyName.MAX_PAGE_RESULTS);
 
 	
 	@Override
