+++ src/main/java/com/julvez/pfc/teachonsnap/controller/common/PagerController.java
@@ -32,7 +32,7 @@ public abstract class PagerController extends CommonController {
 	protected StringManager stringManager = StringManagerFactory.getManager();
 	private PropertyManager properties = PropertyManagerFactory.getManager();
 	
-	protected final int MAX_RESULTS_PAGE = properties.getNumericProperty(LessonPropertyName.MAX_PAGE_RESULTS);
+	protected final int MAX_RESULTS_PAGE = (int)properties.getNumericProperty(LessonPropertyName.MAX_PAGE_RESULTS);
 
 	@Override
 	protected boolean isPrivateZone() {
