+++ src/main/java/com/julvez/pfc/teachonsnap/controller/common/admin/FollowAuthorController.java
@@ -36,7 +36,7 @@ public class FollowAuthorController extends AdminController {
 
 	private StringManager stringManager = StringManagerFactory.getManager();
 	
-	private final int MAX_RESULTS_PAGE = properties.getNumericProperty(UserPropertyName.MAX_PAGE_RESULTS);
+	private final int MAX_RESULTS_PAGE = (int)properties.getNumericProperty(UserPropertyName.MAX_PAGE_RESULTS);
 
 	
 	@Override
