+++ src/main/java/com/julvez/pfc/teachonsnap/controller/common/admin/GroupsController.java
@@ -34,7 +34,7 @@ public class GroupsController extends AdminController {
 
 	private StringManager stringManager = StringManagerFactory.getManager();
 	
-	private final int MAX_RESULTS_PAGE = properties.getNumericProperty(UserPropertyName.MAX_PAGE_RESULTS);
+	private final int MAX_RESULTS_PAGE = (int)properties.getNumericProperty(UserPropertyName.MAX_PAGE_RESULTS);
 
 	
 	@Override
